Add employee search

+++ src/services/department.service.ts
@@ -5,25 +5,26 @@ import { handleServiceError } from '../utils/error.helper';
 import { toDepartmentResponse } from '../utils/mapper'; 
 import { addDepartmentSchema } from '../validation/department/addDepartment.schema';
 import { deleteDepartmentSchema } from '../validation/department/deleteDepartment.schema'; 
-import { getDepartmentByIdSchema } from '../validation/department/getDepartmentById.schema';
-import { getDepartmentByNameSchema } from '../validation/department/getDepartmentByName.schema';
+import { getDepartmentByIdSchema } from '../validation/department/getDepartmentById.schema'; 
 import { updateDepartmentSchema } from '../validation/department/updateDepartment.schema'; 
 import { validate } from '../validation/validate';
 
-export async function getDepartments(): Promise<ServiceResult<IDepartment[]>> {
+//Service export functions
+
+export async function getDepartmentsAsync(): Promise<ServiceResult<IDepartment[]>> {
   try { 
     const departments = await DepartmentModel.find().exec(); 
     return { success: true, data: departments };
-  } catch (err: any) {
-    return { success: false, error: ['Unexpected error: ' + err.message] };
+  } catch (err: unknown) {
+    return handleServiceError(err); 
   }
 } 
   
-export async function getDepartmentById(id: unknown): Promise<ServiceResult<IDepartment>> {
+export async function getDepartmentByIdAsync(id: unknown): Promise<ServiceResult<IDepartment>> {
    
   const validationResult = await validate(getDepartmentByIdSchema, {id});  
   if (!validationResult.success) { 
-    return validationResult;
+    return { success: false, code: 400, error: validationResult.error }
   }    
 
   try {
@@ -32,62 +33,55 @@ export async function getDepartmentById(id: unknown): Promise<ServiceResult<IDep
 
     const department = await DepartmentModel.findById(validId).exec();
     if (!department) {
-      return { success: false, error: ['Department not found'] };
-    }
+      return { success: false, error: ['Department not found'], code: 404 };
+    } 
 
     return { success: true, data: department };
   } 
-  catch (err: any) { 
+  catch (err: unknown) { 
     return handleServiceError(err);
   }
 } 
-
-export async function getDepartmentByName(name: string): Promise<ServiceResult<IDepartment | null>> { 
-
-  const validationResult = await validate(getDepartmentByNameSchema, name);  
-  if (!validationResult.success) {
-    return validationResult;
-  }
-  
-  try {
-    const { name: validName } = validationResult.data;
-    const department = await DepartmentModel.findOne({ name: validName }).exec();
-    return { success: true, data: department }; 
-  } 
-  catch (err: any) { 
-    return handleServiceError(err);
-  }
-}
-
-export async function addDepartment(data: unknown): Promise<ServiceResult<DepartmentResponse>> {
+ 
+export async function addDepartmentAsync(data: unknown): Promise<ServiceResult<DepartmentResponse>> {
     
   const validationResult = await validate(addDepartmentSchema, data);  
   if (!validationResult.success) {
-    return validationResult;
+    return { success: false, code: 400, error: validationResult.error }
   }  
   const validData = validationResult.data;
  
   try {
+
+    if ((await departmentNameExistsAsync(validData.name))) {
+      return {success: false, code: 400, error: ['Department name exists already']};
+    }   
+
     const department = new DepartmentModel(validData);
     const saved = await department.save();  
     return { success: true, data: toDepartmentResponse(saved) };
   } 
-  catch (err: any) {  
+  catch (err: unknown) {  
     return handleServiceError(err); 
   }
 } 
 
-export async function updateDepartment(id: string, name: string): Promise<ServiceResult<UpdatedDepartmentResponse>> {
+export async function updateDepartmentAsync(id: string, name: string): Promise<ServiceResult<UpdatedDepartmentResponse>> {
       
   const validationResult = await validate(updateDepartmentSchema, { id, name });  
   if (!validationResult.success) {
-    return validationResult;
+    return { success: false, code: 400, error: validationResult.error }
   }   
   
   try {
 
     const { id: validId, name: validName } = validationResult.data;
 
+    const existingDepartment = await DepartmentModel.findOne({ name: validName }); 
+    if (existingDepartment && existingDepartment.id !== validId) {
+      return {success: false, code: 404, error: ['Department name exists already']};
+    }
+ 
     const updatedDepartment = await DepartmentModel.findByIdAndUpdate(
       validId,
       { $set: { name: validName } },
@@ -95,32 +89,46 @@ export async function updateDepartment(id: string, name: string): Promise<Servic
     );
 
     if (!updatedDepartment) {
-      return { success: false, error: ['Department not found'] };
+      return { success: false, error: ['Department not found'], code: 400 };
     }
 
     const updatedDepartmentResponse: UpdatedDepartmentResponse = { message: "Department updated successfully", departmentId: updatedDepartment.id }; 
     return { success: true, data: updatedDepartmentResponse }; 
   } 
-  catch (err: any) { 
+  catch (err: unknown) { 
     return handleServiceError(err);
   }
 } 
 
-export async function deleteDepartment(id: string): Promise<ServiceResult<IDepartment | null>> {
+export async function deleteDepartmentAsync(id: string): Promise<ServiceResult<IDepartment | null>> {
   
   const validationResult = await validate(deleteDepartmentSchema, { id });  
   if (!validationResult.success) {
-    return validationResult;
+    return { success: false, code: 400, error: validationResult.error }
   } 
 
   try {
 
     const { id: validId } = validationResult.data;
 
+    if (!(await departmentExistsAsync(validId))) {
+      return {success: false, code: 404, error: ['Department not found']};
+    }   
+
     await DepartmentModel.findByIdAndDelete(validId);
     return { success: true, data: null };
   } 
-  catch (err: any) { 
+  catch (err: unknown) { 
     return handleServiceError(err);
   }
+}
+
+export async function departmentExistsAsync(id: string): Promise<boolean> {
+  const exists = await DepartmentModel.exists({ _id: id }).exec();
+  return !!exists;
+}
+
+export async function departmentNameExistsAsync(departmentName: string): Promise<boolean> {
+  const exists = await DepartmentModel.exists({ 'name': departmentName }).exec();
+  return !!exists;
 }
